+++ frontend/new-react-vite/index.html
@@ -7,13 +7,6 @@
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <style>
-      h1 {
-        font-family: "Rajdhani";
-        color: yellow;
-        font-weight: 300;
-      }
-    </style>
     <title>Vite + React + TS</title>
   </head>
   <body>
