+++ frontend/new-react-vite/src/App.css
@@ -4,7 +4,7 @@
   padding: 2rem;
   text-align: left;
 } */
-.a {
+/* .a {
   
 }
 .rajdhani-light {
@@ -39,4 +39,4 @@
 
 body {
   font-family: 'Rajdhani', serif;
-}
+} */
