+++ frontend/new-react-vite/src/App.tsx
@@ -1,20 +1,22 @@
 import ReactDOM from "react-dom/client";
 import { BrowserRouter, Routes, Route } from "react-router";
-import Layout from "./pages/Layout";
-import Login from "./pages/login";
-import Marketplace from "./pages/Marketplace";
-import Inventory from "./pages/Inventory";
-import NoPage from "./pages/NoPage";
+import Layout from "./components/pages/Layout";
+import Login from "./components/pages/login";
+import Marketplace from "./components/pages/Marketplace";
+import Inventory from "./components/pages/Inventory";
+import NoPage from "./components/pages/NoPage";
 //import reactLogo from './assets/react.svg'
 //import viteLogo from '/vite.svg'
 import './App.css'
+
 const App = () => {
   return (
 
     <div style={{
         color: "white",
-
+        
     }}>
+      
         <BrowserRouter>
             <Routes>
                 <Route path="/" element={<Layout />}>
