+++ frontend/new-react-vite/src/components/cyberpunk-css-main/demo/helper.js
