+++ frontend/new-react-vite/src/components/cyberpunk-css-main/demo/img/favicon.png
