+++ frontend/new-react-vite/src/components/pages/CRTScreen.tsx
