+++ frontend/new-react-vite/src/components/pages/pageCss/layout.css
@@ -1,3 +1,3 @@
-.navBar {
+/* .navBar {
     
-}
+} */
