+++ frontend/new-react-vite/src/components/pages/pageCss/login.css
@@ -5,7 +5,10 @@
 }
 
 body {
-    margin: right;
+  margin: 0;
+  display: flex;
+  flex-direction: column;
+  height: 100vh; 
 }
