+++ frontend/new-react-vite/src/index.css
@@ -25,7 +25,7 @@ a:hover {
 body {
   margin: left;
   display: flex;
-  /* place-items: center; */
+
   min-width: 320px;
   min-height: 100vh;
 }
